+++ src/decorators/services/index.ts
@@ -1,4 +1,6 @@
 
 
 export * from './injectable.decorator';
+export * from './meta.decorator';
+export * from './name.decorator';
 export * from './service.decorator';