+++ src/decorators/services/name.decorator.ts
@@ -1,0 +1,6 @@
+import { CtorT } from '../../core/ctor';
+import { Meta } from './meta.decorator';
+
+export function Name(value: string) {
+    return Meta('name', value);
+}