+++ src/metadata/server/provide.ts
@@ -1,6 +1,8 @@
 import { CtorT } from '../../core/ctor';
 
-export type ProvideMetadataCallbackSource<T> = { provide: CtorT<T>, useCallback: { (...args: any[]): T }, cache?: boolean, deps?: CtorT<any>[] };
+export type DependencySource = CtorT<any> | string;
+
+export type ProvideMetadataCallbackSource<T> = { provide: CtorT<T>, useCallback: { (...args: any[]): T }, cache?: boolean, deps?: DependencySource[] };
 export type ProvideMetadataValueSource<T> = { provide: CtorT<T>, useValue: T };
 export type ProvideMetadataClassSource<T, TImpl extends T> = { provide: CtorT<T>, useClass: CtorT<TImpl> };