+++ src/metadata/services/metadata.ts
@@ -1,0 +1,3 @@
+
+
+export const MetadataMetadataSym = Symbol.for('MetadataMetadata');